add done button to input to dismiss numpad. add swap button. fix bug where doubles were not parsed by model decoder.

+++ currency-converter-andela/Models/ConvertResult.swift
@@ -20,6 +20,6 @@ struct ConvertResultInfo: Codable {
   let timestamp: Int
 }
 struct ConvertResultQuery: Codable {
-  let amount: Int
+  let amount: Double
   let from, to: String
 }
